+++ terra_geocrud/jsoneditor.py
@@ -1,11 +1,11 @@
-from django_json_widget.widgets import JSONEditorWidget
+from django_json_widget.widgets import JSONEditorWidget as BaseJSONEditorWidget
 
 
-class JSONEditorWidget(JSONEditorWidget):
+class JSONEditorWidget(BaseJSONEditorWidget):
     """
     Override media infos ith compliant django style
     Wait for https://github.com/jmrivas86/django-json-widget/pull/23 merge
     """
-    class Media(JSONEditorWidget.Media):
+    class Media(BaseJSONEditorWidget.Media):
         css = {'all': ('dist/jsoneditor.min.css',)}
         js = ('dist/jsoneditor.min.js',)